+++ src/components/DeleteAppForm/DeleteAppForm.jsx
@@ -1,8 +1,23 @@
 import { Alert, Row, Col } from 'react-bootstrap';
+import { useDispatch } from 'react-redux';
+import { useHistory } from 'react-router';
+import { useAppContext } from '../../contexts/App.context';
 
 import DeleteAppModal from '../DeleteAppModal/DeleteAppModal';
 
-const DeleteAppForm = (props) => {
+import { deleteCurrentApp } from '../../ducks/apps/current';
+
+const DeleteAppForm = () => {
+  const history = useHistory();
+  const dispatch = useDispatch();
+
+  const { appId } = useAppContext();
+
+  const onDelete = async () => {
+    await dispatch(deleteCurrentApp(appId));
+    history.push('/applications');
+  };
+
   return (
     <>
       <h3 className={'my-3'}>Danger zone</h3>
@@ -13,7 +28,7 @@ const DeleteAppForm = (props) => {
             <p>All your apps using this client will stop working.</p>
           </Col>
           <Col xs='auto' className={'align-self-center'}>
-            <DeleteAppModal onDelete={props.onDelete} />
+            <DeleteAppModal onDelete={onDelete} />
           </Col>
         </Row>
       </Alert>
