+++ src/pages/ApplicationsPage/ApplicationsPage.jsx
@@ -1,26 +1,13 @@
 import classNames from 'classnames';
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { Container, Row, Col } from 'react-bootstrap';
 
 import AppList from '../../components/AppList/AppList';
 import CreateAppModal from '../../components/CreateAppModal/CreateAppModal';
-import { addApp, clearAllApps, loadAllApps } from '../../ducks/apps/all';
-
-import selector from './ApplicationsPage.selector';
+import { addApp } from '../../ducks/apps/all';
 
 const ApplicationsPage = () => {
   const dispatch = useDispatch();
-
-  const apps = useSelector(selector);
-
-  useEffect(() => {
-    dispatch(loadAllApps());
-
-    return () => {
-      dispatch(clearAllApps());
-    };
-  }, [dispatch]);
 
   const onAdd = (values) => {
     dispatch(addApp(values));
@@ -28,7 +15,7 @@ const ApplicationsPage = () => {
 
   return (
     <Container>
-      <Row className={classNames('mb-5', 'align-items-center')}>
+      <Row className={classNames('mb-4', 'align-items-center')}>
         <Col>
           <h2>Applications</h2>
         </Col>
@@ -37,8 +24,8 @@ const ApplicationsPage = () => {
         </Col>
       </Row>
       <Row>
-        <Col>
-          <AppList list={apps.all} loading={apps.loading} error={apps.error} />
+        <Col className='mb-4'>
+          <AppList />
         </Col>
       </Row>
     </Container>
